return float, not decimal from postgres

+++ api/core/data/processing/common.py
@@ -58,8 +58,8 @@ class Common:
         sql = """
             select
                 p.id as sampling_point_id,
-                extract(epoch from p.from_time) as from_time,
-                extract(epoch from p.to_time) as to_time,
+                extract(epoch from p.from_time)::float as from_time,
+                extract(epoch from p.to_time)::float as to_time,
                 case when t.timestep = 1 and t.notation != 's' then -1 else t.timestep end as timestep,
                 case when cs.id is NULL then False else True end as is_calculated
             from
